move utils to common (#278)
