+++ src/common/util/boolAlgebra.ts
