+++ src/common/util/config.ts
@@ -1,3 +1,4 @@
+let location = window.location;
 let port = location.port;
 
 // development
@@ -10,7 +11,7 @@ export const URL = location.protocol + '//' + location.hostname + port;
 export const API_URL = URL + "/api";
 
 let wsProtocol = "ws:";
-if(location.protocol === "https:")
+if (location.protocol === "https:")
     wsProtocol = "wss:";
 export const SOCKET_URL = wsProtocol + "//" + location.hostname + port + "/ws/";
