+++ src/common/util/random.ts
