+++ src/common/util/reactions/enum.ts
@@ -1,0 +1,7 @@
+export enum ReactionName {
+    HAPPY = "HAPPY",
+    SAD = "SAD",
+    UP = "UP",
+    DOWN = "DOWN",
+    HEART = "HEART"
+}
