+++ src/lecturer/components/timestampView/TimestampTable.tsx
@@ -1,8 +1,8 @@
 import { Paper, makeStyles, Button, useTheme } from "@material-ui/core";
-import { timestampMock } from "../../util/mockData";
 import {TimestampRow} from "./TimestampRow";
 import { useState } from "react";
 import Timeline from '@material-ui/lab/Timeline';
+import { timestampMock } from "../../../common/util/mockData";
 export function TimestampTable() {
     const [filterType, setFilterType] = useState("QuestionType");
     const theme = useTheme();
